requirements.txt
flask
feedparser
tensorflow
numpy==1.16.1
mysqlclient
mecab-python3